Refactor User module to separate Admin and Privilege classes in a separate module.
